+++ assets/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v0.1.2 - 2025-11-29 UTC
 
 - Fix the magnet lever on the ship wall not reflecting the magnet state when toggled via keybind.
+- Disable keybind actions when the player is using menu, terminal, chat, or dead.
 
 ## v0.1.1 - 2025-11-29 UTC